experimental/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
google-generativeai
google-api-python-client
python-dotenv
pandas
numpy
